Enable video transcode from media picker by default.
