+++ assets/src/edit-story/components/library/panes/media/local/mediaPane.js
@@ -148,9 +148,6 @@ function MediaPane(props) {
   );
 
   const { showSnackbar } = useSnackbar();
-  const enableMediaPickerVideoOptimization = useFeature(
-    'enableMediaPickerVideoOptimization'
-  );
 
   const {
     allowedTranscodableMimeTypes,
@@ -164,11 +161,7 @@ function MediaPane(props) {
   const { isFeatureEnabled, isTranscodingEnabled } = useFFmpeg();
 
   const allowedMimeTypes = useMemo(() => {
-    if (
-      enableMediaPickerVideoOptimization &&
-      isFeatureEnabled &&
-      isTranscodingEnabled
-    ) {
+    if (isFeatureEnabled && isTranscodingEnabled) {
       return [
         ...allowedTranscodableMimeTypes,
         ...allowedImageMimeTypes,
@@ -181,7 +174,6 @@ function MediaPane(props) {
     allowedVideoMimeTypes,
     isFeatureEnabled,
     isTranscodingEnabled,
-    enableMediaPickerVideoOptimization,
     allowedTranscodableMimeTypes,
   ]);
 
@@ -210,7 +202,6 @@ function MediaPane(props) {
     const resource = getResourceFromMediaPicker(mediaPickerEl);
     try {
       if (
-        enableMediaPickerVideoOptimization &&
         isFeatureEnabled &&
         isTranscodingEnabled &&
         transcodableMimeTypes.includes(resource.mimeType)
