+++ packages/e2e-test-utils/src/toggleVideoOptimization.js
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * Internal dependencies
+ */
+import visitSettings from './visitSettings';
+
+async function toggleVideoOptimization() {
+  await visitSettings();
+  const selector = '[data-testid="media-optimization-settings-checkbox"]';
+  await page.waitForSelector(selector);
+  // Clicking will only act on the first element.
+  await expect(page).toClick(selector);
+  // Await REST API request.
+  await page.waitForResponse((response) =>
+    response.url().includes('web-stories/v1/users/me')
+  );
+}
+export default toggleVideoOptimization;
