+++ packages/e2e-tests/src/specs/editor/editor.js
@@ -20,22 +20,10 @@
 import percySnapshot from '@percy/puppeteer';
 import {
   createNewStory,
-  visitSettings,
+  toggleVideoOptimization,
   previewStory,
   withRTL,
 } from '@web-stories-wp/e2e-test-utils';
-
-async function toggleVideoOptimization() {
-  await visitSettings();
-  const selector = '[data-testid="media-optimization-settings-checkbox"]';
-  await page.waitForSelector(selector);
-  // Clicking will only act on the first element.
-  await expect(page).toClick(selector);
-  // Await REST API request.
-  await page.waitForResponse((response) =>
-    response.url().includes('web-stories/v1/users/me')
-  );
-}
 
 describe('Story Editor', () => {
   it('should be able to create a blank story', async () => {
